move thread handler
